+++ pipelinewise/cli/config.py
@@ -476,7 +476,7 @@ class Config:
                 'parallelism': tap.get('parallelism', 0),
                 'parallelism_max': tap.get('parallelism_max', 4),
                 'hard_delete': tap.get('hard_delete', True),
-                'flush_all_streams': tap.get('flush_all_streams', False),
+                'flush_all_streams': tap.get('flush_all_streams', True),
                 'primary_key_required': tap.get('primary_key_required', True),
                 'default_target_schema': tap.get('default_target_schema'),
                 'default_target_schema_select_permissions': tap.get(